requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
discord-py-interactions==5.11.0
discord-typings==0.8.0
dnspython==2.6.1
emoji==2.11.0
frozenlist==1.4.1
idna==3.6
Levenshtein==0.25.0
motor==3.4.0
multidict==6.0.5
pymongo==4.6.3
python-dotenv==1.0.1
python-Levenshtein==0.25.0
rapidfuzz==3.7.0
tomli==2.0.1
typing_extensions==4.10.0
yarl==1.9.4
